% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\title{Annotate a flow graph}
\usage{
annotate(g, annotations = NULL, nodes = TRUE, ports = TRUE)
}
\arguments{
\item{g}{flow graph to annotate}

\item{annotations}{annotation database or JSON source for annotations
(by default, the Data Science Ontology via remote database)}

\item{nodes}{whether to annotate nodes}

\item{ports}{whether to annotate ports}
}
\value{
A flow graph with annotated nodes and/or ports.
}
\description{
Annotate a flow graph, as created by \code{record}.
}
\details{
Normally, this function need not be called directly. Instead,
call \code{record} with the switch \code{annotate=TRUE}.

This function modifies the flow graph in place.
}
\seealso{
\code{\link{record}}
}
