% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotation_db}
\alias{annotation_db}
\title{Annotation DB}
\description{
Database of class, function, and method annotations.
}
\section{Methods}{

\code{load_documents(docs)} \cr
Load a list of (parsed) JSON documents conforming to the JSON schema for
annotations.

\code{load_json(txt)} \cr
Load documents from a JSON string, URL, or file.
}

