% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{call_info}
\alias{call_info}
\title{Get information about function call}
\usage{
call_info(call, env = rlang::caller_env(), fun = NULL)
}
\description{
A one-stop shop for information about a function being called,
supporting functions (closures and primitives) and methods (S3, S4, R5, R6).
}
