% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{call_type}
\alias{call_type}
\title{Get type of a call}
\usage{
call_type(call)
}
\description{
Get the type ("named", "namespaced", "recursive", or "inlined") 
of a function call.
}
\seealso{
\code{rlang::lang_type_of}
}
