% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{dict}
\alias{dict}
\alias{keys}
\alias{values}
\alias{has_key}
\alias{del}
\alias{get_default}
\title{Dictionary}
\usage{
dict(...)

keys(d)

values(d)

has_key(d, k)

del(d, k)

get_default(d, k, default = NULL)
}
\description{
A dictionary (hash map) data structure.
The keys must be strings and the values can be any R object.
}
\details{
This data type is a thin wrapper around a standard R environment.
It exists because of the packages currently on CRAN, \code{hashmap} does
not allow arbitrary value types and \code{hash} is GPL-ed.
}
