% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{fun_args_match}
\alias{fun_args_match}
\alias{call_args_match}
\title{Match arguments of function call}
\usage{
fun_args_match(fun_args, call_args)

call_args_match(call, env = rlang::caller_env())
}
\description{
A reimplementation of R's algorithm for argument matching,
including partial matching.
}
