% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{fun_package}
\alias{fun_package}
\title{Get package of function}
\usage{
fun_package(f)
}
\description{
In which package is the function defined?
}
\details{
This function reliably determines the definining package of a
function object, including for primitive functions, anonymous functions,
and methods.
}
