% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph}
\alias{graph}
\alias{is_directed}
\alias{nodes}
\alias{edges}
\alias{nnodes}
\alias{nedges}
\alias{add_node}
\alias{add_nodes}
\alias{has_node}
\alias{rem_node}
\alias{add_edge}
\alias{has_edge}
\alias{rem_edge}
\alias{neighbors}
\alias{successors}
\alias{predecessors}
\title{Graphs}
\usage{
graph(...)

is_directed(g)

nodes(g)

edges(g, ...)

nnodes(g)

nedges(g, ...)

add_node(g, node, ...)

add_nodes(g, nodes)

has_node(g, node)

rem_node(g, node)

add_edge(g, src, tgt, ...)

has_edge(g, src, tgt)

rem_edge(g, src, tgt, ...)

neighbors(g, node)

successors(g, node)

predecessors(g, node)
}
\description{
Generic interface for graphs as S3 objects
(directed or undirected, simple or multi-edged).
}
\details{
Unlike most R objects, graphs are mutable and have
pass-by-reference semantics.
}
