% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect_call}
\alias{inspect_call}
\title{Inspect call}
\usage{
inspect_call(call, env = rlang::caller_env(), fun = NULL)
}
\value{
A named list with items possibly including:
\itemize{
\item \code{name}: Name of function
\item \code{package}: Package where function is defined
\item \code{system}: Class system used by function, if it's a method
}
}
\description{
A convenience method for inspecting function calls, which
supports functions (closures and primitives) and methods (S3, S4, R5, R6).
}
