% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect_obj}
\alias{inspect_obj}
\title{Inspect object}
\usage{
inspect_obj(x)
}
\value{
A named list with items including:
\itemize{
\item \code{class}: Classes of object.
\item \code{system}: Class system used by object.
}
}
\description{
A convenience function for inspecting function objects.
}
\seealso{
\code{\link{inspect_call}}
}
