% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{match_call}
\alias{match_call}
\title{Match arguments}
\usage{
match_call(call, env = rlang::caller_env(), fun = NULL)
}
\description{
Match arguments of call to argument names of called function.
}
\details{
Mostly a wrapper around \code{match.call}, except in the case of
primitive functions, which \code{match.call} doesn't support.
}
\seealso{
\code{\link{match.call}}
}
