% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigraph.R
\name{multigraph}
\alias{multigraph}
\alias{multiedge}
\title{Multigraphs}
\usage{
multigraph(data = dict())

multiedge(src, tgt, ind)
}
\description{
A directed, attributed multigraph.
The internal data structure uses adjacency lists for successors and predecessors.
}
\details{
TODO
}
