% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{obj_package}
\alias{obj_package}
\title{Get package of object}
\usage{
obj_package(x)
}
\description{
In which package is the object's class defined?
}
\details{
This function tries to determine the defining package of the
object's class. In general, this is not possible because R has an informal
class system.

Currently, this function works reliably on S4 and R5 instances, but not on
S3 or R6 instances. When it fails, it returns \code{NULL}.
}
\seealso{
\code{\link{fun_package}}
}
