% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordered_dict.R
\name{ordered_dict}
\alias{ordered_dict}
\alias{as_ordered_dict}
\title{Ordered dictionary}
\usage{
ordered_dict(...)

as_ordered_dict(x)
}
\description{
Data structure for ordered dictionaries,
ala \code{OrderedDict} in Python or Julia.
}
\details{
Like ordinary dictionaries (\code{dict}) and unlike most R objects,
ordered dictionaries are mutable. The implementation combines a dictionary
with a circular doubly linked list, exactly as in Python's
\code{OrderedDict}.
}
\seealso{
\code{\link{dict}}
}
