% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphml.R
\name{read_graphml}
\alias{read_graphml}
\title{Read graph from GraphML}
\usage{
read_graphml(xml, graph = NULL)
}
\arguments{
\item{xml}{XML document or input to \code{xml2::read_xml}}
}
\description{
Read graph from GraphML
}
