% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record}
\alias{record}
\alias{record_file}
\alias{record_expr}
\title{Record as flow graph}
\usage{
record(x, env = rlang::caller_env(), ...)

record_file(x, env = rlang::caller_env(), ...)

record_expr(x, env = rlang::caller_env(), out = NULL, cwd = NULL,
  annotate = FALSE, db = NULL, data = annotate, node_data = data,
  port_data = data, port_values = FALSE)
}
\arguments{
\item{x}{(unevaluated) expression}

\item{env}{evaluation environment}

\item{out}{file path or connection in which to write flow graph as GraphML}

\item{cwd}{working directory during evaluation (by default, the current
working directory)}

\item{annotate}{whether to annotate the flow graph}

\item{db}{annotation database or JSON source for annotations, if annotation
is enabled (by default, the Data Science Ontology via remote database)}

\item{data}{convenience switch to store metadata (on nodes and ports)}

\item{node_data}{whether to store function metadata as node data}

\item{port_data}{whether to store object metadata as port data}

\item{port_values}{whether to store object values as port data}
}
\value{
A flow graph, of class \code{wiring_diagram}, for the evaluation.
}
\description{
Record the evaluation of an R expression as a flow graph.
}
