% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{record_in_ast}
\alias{record_in_ast}
\title{Add recording to R program by transforming the AST.}
\usage{
record_in_ast(expr, context = NULL)
}
\description{
Add recording to R program by transforming the AST.
}
