% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_db.R
\name{remote_annotation_db}
\alias{remote_annotation_db}
\title{Remote annotation DB}
\description{
Annotation database that fetches annotations from a remote 
CouchDB database.
}
\section{Methods}{

\code{load_all_packages()} \cr
Load annotations for all R packages.

\code{load_package(package)} \cr
Load annotations for a specific R package.
}

\seealso{
\code{\link{annotation_db}}
}
