% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiring_diagram.R
\name{wiring_diagram}
\alias{wiring_diagram}
\alias{input_node}
\alias{output_node}
\alias{add_node.wiring_diagram}
\alias{input_ports}
\alias{output_ports}
\alias{add_edge.wiring_diagram}
\alias{source_port}
\alias{target_port}
\title{Wiring diagrams}
\usage{
wiring_diagram(in_ports = list(), out_ports = list(), data = list())

input_node(g)

output_node(g)

\method{add_node}{wiring_diagram}(g, node, in_ports, out_ports, data = list())

input_ports(g, node = NULL)

output_ports(g, node = NULL)

\method{add_edge}{wiring_diagram}(g, src, tgt, src_port, tgt_port,
  data = list())

source_port(g, src, tgt, ind)

target_port(g, src, tgt, ind)
}
\description{
Wiring diagrams (aka string diagrams) with arbitrary node/edge
data.
}
\details{
The implementation is a simple wrapper around \code{multigraph}.
}
\seealso{
\code{\link{multigraph}}, \code{\link{graph}}
}
