% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiring_diagram.R
\name{wiring_diagram_data}
\alias{wiring_diagram_data}
\alias{input_port_data}
\alias{input_port_data<-}
\alias{input_port_attr}
\alias{input_port_attr<-}
\alias{output_port_data}
\alias{output_port_data<-}
\alias{output_port_attr}
\alias{output_port_attr<-}
\title{Wiring diagram data}
\usage{
input_port_data(g, node, port)

input_port_data(g, node, port) <- value

input_port_attr(g, node, port, key)

input_port_attr(g, node, port, key) <- value

output_port_data(g, node, port)

output_port_data(g, node, port) <- value

output_port_attr(g, node, port, key)

output_port_attr(g, node, port, key) <- value
}
\description{
In addition to the usual graph/node/edge data, wiring diagrams
allow arbitrary data on input and output ports.
}
\seealso{
\code{\link{graph_data}}
}
