% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{dict}
\alias{dict}
\alias{as_dict}
\alias{keys}
\alias{values}
\alias{has_key}
\alias{del}
\alias{get_default}
\alias{set_default}
\title{Dictionary}
\usage{
dict(...)

as_dict(x)

keys(d)

values(d)

has_key(d, k)

del(d, k)

get_default(d, k, default = NULL)

set_default(d, k, default = NULL)
}
\description{
A dictionary (hash map) data structure.
The keys must be strings and the values can be any R object.
}
\details{
The \code{dict} data type is a thin wrapper around an R environment.
Unlike most R types, it is mutable and has pass-by-reference semantics.

The \code{dict} interface is also implemented for lists, which can serve as
immutable dictionaries (with linear time lookup).
}
