% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigraph.R
\name{multigraph}
\alias{multigraph}
\alias{multiedge}
\title{Multigraphs}
\usage{
multigraph(data = list())

multiedge(src, tgt, ind)
}
\description{
A directed, attributed multigraph.
}
\details{
The internal data structure uses adjacency lists for successors
and predecessors. It is modeled on the \code{MultiDiGraph} class in the
Python library NetworkX.
}
