% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordered_dict.R
\name{ordered_dict}
\alias{ordered_dict}
\title{Ordered dictionary}
\usage{
ordered_dict()
}
\description{
An ordered dictionary data structure, ala \code{OrderedDict}
in Python or Julia.
}
\details{
TODO
}
