% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record}
\alias{record}
\title{Record as flow graph}
\usage{
record(x, env = rlang::caller_env(), annotate = FALSE, db = NULL,
  data = annotate, node_data = data, port_data = data, values = data)
}
\arguments{
\item{x}{(unevaluated) expression to evaluate and record}

\item{env}{evaluation environment}

\item{annotate}{whether to annotate the flow graph}

\item{db}{annotation database, if annotation is enabled
(by default, the Data Science Ontology)}

\item{data}{convenience switch for data storage 
(node data, port data, values)}

\item{node_data}{whether to store function metadata as node data}

\item{port_data}{whether to store object metadata as port data}

\item{values}{whether to store object values as port data}
}
\value{
A flow graph, of class \code{wiring_diagram}, for the evaluation.
}
\description{
Record the evaluation of an R expression as a flow graph.
}
